+++ src/textual/_node_list.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import sys
 import weakref
 from operator import attrgetter
-from typing import TYPE_CHECKING, Any, Callable, Iterable, Iterator, Sequence, overload
+from typing import TYPE_CHECKING, Any, Callable, Iterator, Sequence, overload
 
 import rich.repr
 
@@ -12,6 +12,10 @@ if TYPE_CHECKING:
 
     from textual.dom import DOMNode
     from textual.widget import Widget
+
+
+_display_getter = attrgetter("display")
+_visible_getter = attrgetter("visible")
 
 
 class DuplicateIds(Exception):
@@ -41,6 +45,8 @@ class NodeList(Sequence["Widget"]):
         # The nodes in the list
         self._nodes: list[Widget] = []
         self._nodes_set: set[Widget] = set()
+        self._displayed_nodes: tuple[int, list[Widget]] = (-1, [])
+        self._displayed_visible_nodes: tuple[int, list[Widget]] = (-1, [])
 
         # We cache widgets by their IDs too for a quick lookup
         # Note that only widgets with IDs are cached like this, so
@@ -69,8 +75,6 @@ class NodeList(Sequence["Widget"]):
         """Mark the nodes as having been updated."""
         self._updates += 1
         node = None if self._parent is None else self._parent()
-        if node is None:
-            return
         while node is not None and (node := node._parent) is not None:
             node._nodes._updates += 1
 
@@ -187,18 +191,29 @@ class NodeList(Sequence["Widget"]):
         return reversed(self._nodes)
 
     @property
-    def displayed(self) -> Iterable[Widget]:
+    def displayed(self) -> Sequence[Widget]:
         """Just the nodes where `display==True`."""
-        for node in self._nodes:
-            if node.display:
-                yield node
+        if self._displayed_nodes[0] != self._updates:
+            self._displayed_nodes = (
+                self._updates,
+                list(filter(_display_getter, self._nodes)),
+            )
+        return self._displayed_nodes[1]
 
     @property
-    def displayed_reverse(self) -> Iterable[Widget]:
+    def displayed_and_visible(self) -> Sequence[Widget]:
+        """Nodes with both `display==True` and `visible==True`."""
+        if self._displayed_visible_nodes[0] != self._updates:
+            self._displayed_nodes = (
+                self._updates,
+                list(filter(_visible_getter, self.displayed)),
+            )
+        return self._displayed_nodes[1]
+
+    @property
+    def displayed_reverse(self) -> Iterator[Widget]:
         """Just the nodes where `display==True`, in reverse order."""
-        for node in reversed(self._nodes):
-            if node.display:
-                yield node
+        return filter(_display_getter, reversed(self._nodes))
 
     if TYPE_CHECKING:
 
@@ -211,9 +226,11 @@ class NodeList(Sequence["Widget"]):
     def __getitem__(self, index: int | slice) -> Widget | list[Widget]:
         return self._nodes[index]
 
-    def __getattr__(self, key: str) -> object:
-        if key in {"clear", "append", "pop", "insert", "remove", "extend"}:
-            raise ReadOnlyError(
-                "Widget.children is read-only: use Widget.mount(...) or Widget.remove(...) to add or remove widgets"
-            )
-        raise AttributeError(key)
+    if not TYPE_CHECKING:
+        # This confused the type checker for some reason
+        def __getattr__(self, key: str) -> object:
+            if key in {"clear", "append", "pop", "insert", "remove", "extend"}:
+                raise ReadOnlyError(
+                    "Widget.children is read-only: use Widget.mount(...) or Widget.remove(...) to add or remove widgets"
+                )
+            raise AttributeError(key)