+++ src/textual/binding.py
@@ -66,7 +66,7 @@ class Binding:
     key_display: str | None = None
     """How the key should be shown in footer.
 
-    If None, the display of the key will use the result of `App.get_key_display`.
+    If `None`, the display of the key will use the result of `App.get_key_display`.
 
     If overridden in a keymap then this value is ignored.
     """
@@ -83,6 +83,16 @@ class Binding:
     """
     system: bool = False
     """Make this binding a system binding, which removes it from the key panel."""
+
+    @dataclass(frozen=True)
+    class Group:
+        """A binding group causes the keys to be grouped under a single description."""
+
+        description: str = ""
+        """Description of the group."""
+
+    group: Group | None = None
+    """Optional binding group (used to group related bindings in the footer)."""
 
     def parse_key(self) -> tuple[list[str], str]:
         """Parse a key into a list of modifiers, and the actual key.
@@ -151,6 +161,7 @@ class Binding:
                     tooltip=binding.tooltip,
                     id=binding.id,
                     system=binding.system,
+                    group=binding.group,
                 )